deluje implementacija

+++ app/src/main/java/com/example/pametnipaketnik/Package.kt
@@ -1,6 +1,6 @@
 package com.example.pametnipaketnik
 
-class Package (val name: String, val x: Int, val y: Int) {
+class Package (val name: String, val x: Double, val y: Double) {
     override fun toString(): String {
         return "Opened(name='$name', x='$x', y='$y')"
     }
